+++ CHANGELOG.md
@@ -16,6 +16,8 @@ FIXES:
 IMPROVEMENTS:
 - Exposed Terraform `refresh` command.
 - Added `refresh` command at the environment and global scope.
+- Added `format` command at the environment and global scope.
+
 ## 0.6.1 (April 8, 2017)
 IMPROVEMENTS:
 - Stack sync no longer sources modules when retrieving state.